% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_link.R
\name{check_link}
\alias{check_link}
\title{check_link}
\usage{
check_link(input_set_number)
}
\arguments{
\item{input_set_number}{Le numéro du set LEGO}
}
\description{
Cette fonction permet de vérifier l'existence du lien de visualisation 3D
sur le site Mecabricks pour un set LEGO. Elle utilise `httr2` pour envoyer une
requête HTTP au site et vérifier l'état de la réponse et renvoie une exception
si le lien est indisponible.
}
