% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_reviews_html.R
\name{create_reviews_html}
\alias{create_reviews_html}
\title{create_reviews_html}
\usage{
create_reviews_html(df_reviews, input_set_number)
}
\arguments{
\item{df_reviews}{Le dataframe des critiques}

\item{input_set_number}{Le numéro du set LEGO}
}
\description{
Cette fonction permet de récupérer l'auteur, la date, le titre, le corps de la
review et enfin la note accordée par l'auteur au set LEGO pour être ensuite
envoyée dans un squelette HTML brut dans l'application.
Si jamais il n'y a pas de critique associée au set, alors la fonction renvoie :
-- Aucun avis disponible pour ce set --.
}
