% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_reviews_html.R
\name{create_stars}
\alias{create_stars}
\title{create_stars}
\usage{
create_stars(note)
}
\arguments{
\item{note}{La note agrégée d'un set LEGO}
}
\description{
Cette fonction permet de transformer la note d'un set en étoiles
et de l'afficher dans l'interface utilisateur. Elle vérifie d'abord
si la note est bien comprise entre 0 et 5 avant de passer au traitement.
}
