% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{custom_theme}
\alias{custom_theme}
\title{Crée un thème personnalisé pour l'application ShinyBricks.}
\usage{
custom_theme()
}
\value{
Un thème personnalisé pour l'application ShinyBricks.
}
\description{
Cette fonction crée un thème personnalisé pour l'application ShinyBricks en utilisant le package bslib.
}
\seealso{
\code{\link{bslib::bs_theme}}
}
