% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_instructions.R
\name{display_instructions}
\alias{display_instructions}
\title{display_instructions}
\usage{
display_instructions(df_instructions, input_set_number)
}
\arguments{
\item{df_instructions}{Le dataframe des instructions}

\item{input_set_number}{Le numéro du set LEGO}
}
\description{
Cette fonction permet d'afficher les instructions disponibles d'un set LEGO
sur le site officiel de LEGO. Elles sont ensuite affichées sous forme d'URL dans
l'application.
}
