% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_3D_link.R
\name{get_3D_link}
\alias{get_3D_link}
\title{get_3D_link}
\usage{
get_3D_link(set_number)
}
\arguments{
\item{set_number}{Le numéro du set LEGO}
}
\description{
Cette fonction permet de créer le lien de visualisation 3D du site Mecabricks
Utilisée seule, elle ne vérifie pas l'existence ou non du lien. En effet,
certains sets n'ont pas de visualisation 3D associée.
}
