% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wordcloud.R
\name{get_wordcloud_adj}
\alias{get_wordcloud_adj}
\title{get_wordcloud_adj}
\usage{
get_wordcloud_adj(df_reviews, input_set_number)
}
\arguments{
\item{df_reviews}{Le dataframe des critiques}

\item{input_set_number}{Le numéro du set LEGO}
}
\description{
Cette fonction permet de créer un nuage de mots des adjectifs les plus fréquents
des avis sur un set. L'adjectif doit faire plus que 3 lettres pour être pris en compte.
}
