% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wordcloud.R
\name{get_wordcloud_all}
\alias{get_wordcloud_all}
\title{get_wordcloud_all}
\usage{
get_wordcloud_all(df_reviews, input_set_number)
}
\arguments{
\item{df_reviews}{Le dataframe des critiques}

\item{input_set_number}{Le numéro du set LEGO}
}
\description{
Cette fonction permet de créer un nuage de mots grâce aux mots les plus fréquents
des avis sur un set. Le mot `set` est exclu car il sera toujours le plus fréquent.
}
