% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grand_total_themes_donut.R
\name{grand_total_themes_donut}
\alias{grand_total_themes_donut}
\title{grand_total_themes_donut}
\usage{
grand_total_themes_donut(df_prices, themes)
}
\arguments{
\item{df_prices}{Le dataframe principal}

\item{themes}{Les thèmes sélectionnés par l'utilisateur dans l'application}
}
\description{
Cette fonction permet de créer un graphique en anneau plotly interactif dont l'objectif
va être de visualiser la part en pourcentage des thèmes sélectionnés par l'utilisateur
dans l'ensemble des thèmes LEGO.
}
