% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_all_parquets.R
\name{import_all_parquets}
\alias{import_all_parquets}
\title{import_all_parquets}
\usage{
import_all_parquets(folder_path = "data/sets")
}
\arguments{
\item{folder_path}{Le chemin du dossier où sont stockés les fichiers parquet}
}
\description{
Cette fonction permet de lister puis d'importer tous les fichiers .parquet contenus dans
le dossier data/sets, c'est à dire l'ensemble des sets par année récupérés
par l'API de Bricklink.
}
