% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricing_hist_plot.R
\name{pricing_hist_plot}
\alias{pricing_hist_plot}
\title{pricing_hist_plot}
\usage{
pricing_hist_plot(df_prices)
}
\arguments{
\item{df_prices}{Le dataframe principal}
}
\description{
Cette fonction permet de créer un histogramme plotly interactif dont l'objectif
va être de visualiser le nombre de sets LEGO compris dans des tranches de 10€
de 0 à 450€.
}
