% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricing_licensed.R
\name{pricing_licensed}
\alias{pricing_licensed}
\title{pricing_licensed}
\usage{
pricing_licensed(df_prices)
}
\arguments{
\item{df_prices}{le dataframe principal}
}
\description{
Cette fonction génère un graphique plotly permettant d'analyser l'évolution des sets
par année selon qu'ils soient sous licence ou non.
}
