% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricing_pieces_plot.R
\name{pricing_pieces_plot}
\alias{pricing_pieces_plot}
\title{Crée un graphique de dispersion des prix par rapport au nombre de pièces.}
\usage{
pricing_pieces_plot(df_prices)
}
\arguments{
\item{df_prices}{Le dataframe contenant les prix de vente au détail et le nombre de pièces de chaque set.}
}
\value{
Un graphique interactif de dispersion des prix par rapport au nombre de pièces.
}
\description{
Cette fonction prend un dataframe contenant les prix de vente au détail et le nombre de pièces de chaque set, puis crée un graphique interactif de dispersion des prix par rapport au nombre de pièces.
}
