% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricing_table.R
\name{pricing_table}
\alias{pricing_table}
\title{Analyse du pricing par thème.}
\usage{
pricing_table(df_prices, theme_input)
}
\arguments{
\item{df_prices}{Le dataframe contenant les prix de vente au détail.}

\item{theme_input}{Le thème pour lequel l'analyse du pricing doit être effectuée.}
}
\value{
Un tableau interactif présentant l'analyse du pricing par thème.
}
\description{
Cette fonction prend un dataframe contenant les prix de vente au détail et effectue une analyse du pricing par thème. Elle calcule le prix minimum, médian, moyen et maximum pour chaque thème.
}
