% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_age.R
\name{set_age_range}
\alias{set_age_range}
\title{Récupère la tranche d'âge recommandée pour un ensemble spécifique.}
\usage{
set_age_range(df_prices, input_set_number)
}
\arguments{
\item{df_prices}{Le dataframe contenant les informations sur les prix des ensembles.}

\item{input_set_number}{Le numéro de set pour lequel récupérer la tranche d'âge recommandée.}
}
\value{
La tranche d'âge recommandée pour l'ensemble spécifié.
}
\description{
Cette fonction prend un dataframe contenant les informations sur les prix des ensembles et récupère la tranche d'âge recommandée pour un ensemble spécifique, identifié par son numéro de set.
}
