% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_figs_number.R
\name{set_figs_number}
\alias{set_figs_number}
\title{Récupère le nombre de figurines pour un ensemble spécifique.}
\usage{
set_figs_number(df_prices, input_set_number)
}
\arguments{
\item{df_prices}{Le dataframe contenant les informations sur les ensembles.}

\item{input_set_number}{Le numéro de set pour lequel récupérer le nombre de figurines.}
}
\value{
Le nombre de figurines pour l'ensemble spécifié.
}
\description{
Cette fonction prend un dataframe contenant les informations sur les prix des ensembles et récupère le nombre de figurines pour un ensemble spécifique, identifié par son numéro de set.
}
