% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_image.R
\name{set_image}
\alias{set_image}
\title{Récupère l'URL de l'image pour un ensemble spécifique.}
\usage{
set_image(df_prices, input_set_number)
}
\arguments{
\item{df_prices}{Le dataframe contenant les informations sur les ensembles.}

\item{input_set_number}{Le numéro de set pour lequel récupérer l'URL de l'image.}
}
\value{
L'URL de l'image pour l'ensemble spécifié.
}
\description{
Cette fonction prend un dataframe contenant les informations sur les ensembles et récupère l'URL de l'image pour un ensemble spécifique, identifié par son numéro de set.
}
