% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_name.R
\name{set_name}
\alias{set_name}
\title{Récupère le nom pour un ensemble spécifique.}
\usage{
set_name(df_prices, input_set_number)
}
\arguments{
\item{df_prices}{Le dataframe contenant les informations sur les ensembles.}

\item{input_set_number}{Le numéro de set pour lequel récupérer le nom.}
}
\value{
Le nom de l'ensemble spécifié.
}
\description{
Cette fonction prend un dataframe contenant les informations sur les ensembles et récupère le nom pour un ensemble spécifique, identifié par son numéro de set.
}
