% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_page.R
\name{set_page}
\alias{set_page}
\title{Récupère l'URL de la page Brickset pour un ensemble spécifique.}
\usage{
set_page(df_prices, input_set_number)
}
\arguments{
\item{df_prices}{Le dataframe contenant les informations sur les ensembles.}

\item{input_set_number}{Le numéro de set pour lequel récupérer l'URL de la page Brickset.}
}
\value{
L'URL de la page Brickset pour l'ensemble spécifié.
}
\description{
Cette fonction prend un dataframe contenant les informations sur les ensembles et récupère l'URL de la page Brickset pour un ensemble spécifique, identifié par son numéro de set.
}
