% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_pieces_number.R
\name{set_pieces_number}
\alias{set_pieces_number}
\title{Récupère le nombre de pièces pour un ensemble spécifique.}
\usage{
set_pieces_number(df_prices, input_set_number)
}
\arguments{
\item{df_prices}{Le dataframe contenant les informations sur les ensembles.}

\item{input_set_number}{Le numéro de set pour lequel récupérer le nombre de pièces.}
}
\value{
Le nombre de pièces pour l'ensemble spécifié.
}
\description{
Cette fonction prend un dataframe contenant les informations sur les ensembles et récupère le nombre de pièces pour un ensemble spécifique, identifié par son numéro de set.
}
