% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_ppp.R
\name{set_ppp}
\alias{set_ppp}
\title{Récupère le prix par pièce pour un ensemble spécifique.}
\usage{
set_ppp(df_prices, input_set_number)
}
\arguments{
\item{df_prices}{Le dataframe contenant les informations sur les ensembles.}

\item{input_set_number}{Le numéro de set pour lequel récupérer le prix par pièce.}
}
\value{
Le prix par pièce pour l'ensemble spécifié.
}
\description{
Cette fonction prend un dataframe contenant les informations sur les ensembles et récupère le prix par pièce pour un ensemble spécifique, identifié par son numéro de set.
}
