% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_prices_lollipops.R
\name{set_prices_lollipops}
\alias{set_prices_lollipops}
\title{Crée un graphique interactif représentant les prix d'un ensemble sous forme de lollipop.}
\usage{
set_prices_lollipops(df_prices, input_set_number)
}
\arguments{
\item{df_prices}{Le dataframe contenant les informations sur les ensembles.}

\item{input_set_number}{Le numéro de set pour lequel générer le graphique.}
}
\value{
Un graphique interactif représentant les prix de l'ensemble spécifié.
}
\description{
Cette fonction prend un dataframe contenant les informations sur les ensembles et crée un graphique interactif représentant les différents prix (prix de détail, prix actuel pour une boîte ouverte, prix actuel pour une boîte scellée) d'un ensemble spécifique, identifié par son numéro de set.
}
