% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_retail_price.R
\name{set_retail_price}
\alias{set_retail_price}
\title{Récupère le prix de détail d'un ensemble spécifique.}
\usage{
set_retail_price(df_prices, input_set_number)
}
\arguments{
\item{df_prices}{Le dataframe contenant les informations sur les ensembles.}

\item{input_set_number}{Le numéro de set pour lequel récupérer le prix de détail.}
}
\value{
Le prix de détail de l'ensemble spécifié.
}
\description{
Cette fonction prend un dataframe contenant les informations sur les ensembles et renvoie le prix de détail d'un ensemble spécifique, identifié par son numéro de set.
}
