% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_value_new.R
\name{set_value_new}
\alias{set_value_new}
\title{Récupère la valeur d'un ensemble neuf spécifique.}
\usage{
set_value_new(df_prices, input_set_number)
}
\arguments{
\item{df_prices}{Le dataframe contenant les informations sur les ensembles.}

\item{input_set_number}{Le numéro de set pour lequel récupérer la valeur neuf.}
}
\value{
La valeur neuf de l'ensemble spécifié.
}
\description{
Cette fonction prend un dataframe contenant les informations sur les ensembles et renvoie la valeur d'un ensemble neuf spécifique, identifié par son numéro de set.
}
