% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_value_used.R
\name{set_value_used}
\alias{set_value_used}
\title{Récupère la valeur d'un ensemble d'occasion spécifique.}
\usage{
set_value_used(df_prices, input_set_number)
}
\arguments{
\item{df_prices}{Le dataframe contenant les informations sur les ensembles.}

\item{input_set_number}{Le numéro de set pour lequel récupérer la valeur d'occasion.}
}
\value{
La valeur d'occasion de l'ensemble spécifié.
}
\description{
Cette fonction prend un dataframe contenant les informations sur les ensembles et renvoie la valeur d'un ensemble d'occasion spécifique, identifié par son numéro de set.
}
