% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_blocktext.R
\name{theme_blocktext}
\alias{theme_blocktext}
\title{Génère un texte descriptif sur un thème LEGO spécifique.}
\usage{
theme_blocktext(df_prices, theme_input)
}
\arguments{
\item{df_prices}{Le dataframe contenant les informations sur les ensembles LEGO.}

\item{theme_input}{Le thème LEGO pour lequel générer le texte descriptif.}
}
\value{
Un texte descriptif sur le thème LEGO spécifié.
}
\description{
Cette fonction prend un dataframe contenant des informations sur les ensembles LEGO et génère un texte descriptif sur un thème LEGO spécifique, en indiquant l'année de lancement, l'année de fin (le cas échéant) et la durée d'existence moyenne des ensembles de ce thème.
}
\seealso{
\code{\link{theme_launch}}, \code{\link{theme_end}}, \code{\link{theme_mean_price}}
}
