% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_end.R
\name{theme_end}
\alias{theme_end}
\title{Détermine l'année de fin d'un thème LEGO spécifique.}
\usage{
theme_end(df_prices, theme_input)
}
\arguments{
\item{df_prices}{Le dataframe contenant les informations sur les ensembles LEGO.}

\item{theme_input}{Le thème LEGO pour lequel déterminer l'année de fin.}
}
\value{
L'année de fin du thème LEGO spécifié.
}
\description{
Cette fonction prend un dataframe contenant des informations sur les ensembles LEGO et détermine l'année de fin d'un thème LEGO spécifique en se basant sur l'année de sortie des ensembles.
}
