% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_group.R
\name{theme_group}
\alias{theme_group}
\title{Détermine le type de groupe pour un thème LEGO spécifique.}
\usage{
theme_group(df_prices, theme_input)
}
\arguments{
\item{df_prices}{Le dataframe contenant les informations sur les ensembles LEGO.}

\item{theme_input}{Le thème LEGO pour lequel déterminer le type de groupe.}
}
\value{
Le type de groupe pour le thème LEGO spécifié.
}
\description{
Cette fonction prend un dataframe contenant des informations sur les ensembles LEGO et détermine le type de groupe pour un thème LEGO spécifique en se basant sur les groupes de thèmes présents dans le dataframe.
}
