% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_launch.R
\name{theme_launch}
\alias{theme_launch}
\title{Détermine l'année de lancement pour un thème LEGO spécifique.}
\usage{
theme_launch(df_prices, theme_input)
}
\arguments{
\item{df_prices}{Le dataframe contenant les informations sur les ensembles LEGO.}

\item{theme_input}{Le thème LEGO pour lequel déterminer l'année de lancement.}
}
\value{
L'année de lancement du thème LEGO spécifié.
}
\description{
Cette fonction prend un dataframe contenant des informations sur les ensembles LEGO et détermine l'année de lancement pour un thème LEGO spécifique en se basant sur les années des ensembles dans le dataframe.
}
