% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_lifespan.R
\name{theme_lifespan}
\alias{theme_lifespan}
\title{Calcule la durée de vie d'un thème LEGO spécifique.}
\usage{
theme_lifespan(df_prices, theme_input)
}
\arguments{
\item{df_prices}{Le dataframe contenant les informations sur les ensembles LEGO.}

\item{theme_input}{Le thème LEGO pour lequel calculer la durée de vie.}
}
\value{
La durée de vie du thème LEGO spécifié.
}
\description{
Cette fonction prend un dataframe contenant des informations sur les ensembles LEGO et calcule la durée de vie d'un thème LEGO spécifique en se basant sur les années de sortie des ensembles dans le dataframe.
}
