% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mean_pieces.R
\name{theme_mean_pieces}
\alias{theme_mean_pieces}
\title{Calcule le nombre moyen de pièces pour un thème LEGO spécifique.}
\usage{
theme_mean_pieces(df_prices, theme_input)
}
\arguments{
\item{df_prices}{Le dataframe contenant les informations sur les ensembles LEGO.}

\item{theme_input}{Le thème LEGO pour lequel calculer le nombre moyen de pièces.}
}
\value{
Le nombre moyen de pièces pour le thème LEGO spécifié.
}
\description{
Cette fonction prend un dataframe contenant des informations sur les ensembles LEGO et calcule le nombre moyen de pièces pour un thème LEGO spécifique.
}
