% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mean_price_new.R
\name{theme_mean_price_new}
\alias{theme_mean_price_new}
\title{Calcule la valeur moyenne des ensembles neufs pour un thème LEGO spécifique.}
\usage{
theme_mean_price_new(df_prices, theme_input)
}
\arguments{
\item{df_prices}{Le dataframe contenant les informations sur les ensembles LEGO.}

\item{theme_input}{Le thème LEGO pour lequel calculer la valeur moyenne des ensembles neufs.}
}
\value{
La valeur moyenne des ensembles neufs pour le thème LEGO spécifié.
}
\description{
Cette fonction prend un dataframe contenant des informations sur les ensembles LEGO et calcule la valeur moyenne des ensembles neufs pour un thème LEGO spécifique.
}
