% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mean_rating.R
\name{theme_mean_rating}
\alias{theme_mean_rating}
\title{Calcule la note moyenne des ensembles pour un thème LEGO spécifique.}
\usage{
theme_mean_rating(df_prices, theme_input)
}
\arguments{
\item{df_prices}{Le dataframe contenant les informations sur les ensembles LEGO.}

\item{theme_input}{Le thème LEGO pour lequel calculer la note moyenne des ensembles.}
}
\value{
La note moyenne des ensembles pour le thème LEGO spécifié.
}
\description{
Cette fonction prend un dataframe contenant des informations sur les ensembles LEGO et calcule la note moyenne des ensembles pour un thème LEGO spécifique.
}
