% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_median_age.R
\name{theme_median_age}
\alias{theme_median_age}
\title{Calcule l'âge médian recommandé pour les ensembles d'un thème LEGO spécifique.}
\usage{
theme_median_age(df_prices, theme_input)
}
\arguments{
\item{df_prices}{Le dataframe contenant les informations sur les ensembles LEGO.}

\item{theme_input}{Le thème LEGO pour lequel calculer l'âge médian recommandé.}
}
\value{
L'âge médian recommandé pour les ensembles du thème LEGO spécifié.
}
\description{
Cette fonction prend un dataframe contenant des informations sur les ensembles LEGO et calcule l'âge médian recommandé pour les ensembles d'un thème LEGO spécifique.
}
