% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_price_multiplier.R
\name{theme_price_multiplier}
\alias{theme_price_multiplier}
\title{Calcule le multiplicateur de prix moyen pour les ensembles neufs par rapport aux ensembles d'occasion d'un thème LEGO spécifique.}
\usage{
theme_price_multiplier(df_prices, theme_input)
}
\arguments{
\item{df_prices}{Le dataframe contenant les informations sur les ensembles LEGO.}

\item{theme_input}{Le thème LEGO pour lequel calculer le multiplicateur de prix.}
}
\value{
Le multiplicateur de prix moyen pour les ensembles neufs par rapport aux ensembles d'occasion du thème LEGO spécifié.
}
\description{
Cette fonction prend un dataframe contenant des informations sur les ensembles LEGO et calcule le multiplicateur de prix moyen pour les ensembles neufs par rapport aux ensembles d'occasion d'un thème LEGO spécifique.
}
\seealso{
\code{\link{theme_mean_price}}, \code{\link{theme_mean_price_new}}
}
