% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_price_range.R
\name{theme_price_range}
\alias{theme_price_range}
\title{Calcule la gamme de prix des ensembles LEGO pour un thème spécifique.}
\usage{
theme_price_range(df_prices, theme_input)
}
\arguments{
\item{df_prices}{Le dataframe contenant les informations sur les ensembles LEGO.}

\item{theme_input}{Le thème LEGO pour lequel calculer la gamme de prix.}
}
\value{
La gamme de prix (min-max) pour le thème LEGO spécifié.
}
\description{
Cette fonction prend un dataframe contenant des informations sur les ensembles LEGO et calcule la gamme de prix (min-max) pour un thème LEGO spécifique.
}
