% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_total_sets.R
\name{theme_total_sets}
\alias{theme_total_sets}
\title{Calcule le nombre total d'ensembles LEGO pour un thème spécifique.}
\usage{
theme_total_sets(df_prices, theme_input)
}
\arguments{
\item{df_prices}{Le dataframe contenant les informations sur les ensembles LEGO.}

\item{theme_input}{Le thème LEGO pour lequel calculer le nombre total d'ensembles.}
}
\value{
Le nombre total d'ensembles pour le thème LEGO spécifié.
}
\description{
Cette fonction prend un dataframe contenant des informations sur les ensembles LEGO et calcule le nombre total d'ensembles pour un thème LEGO spécifique.
}
