% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_10_lineplot.R
\name{top_10_lineplot}
\alias{top_10_lineplot}
\title{Trace un graphique en ligne pour les dix thèmes les plus populaires au fil des années.}
\usage{
top_10_lineplot(df_prices, themes)
}
\arguments{
\item{df_prices}{Le dataframe contenant les données sur les prix des sets.}

\item{themes}{Un vecteur contenant les noms des thèmes à inclure dans le graphique.}
}
\value{
Un graphique en ligne montrant l'évolution du nombre de sets pour les thèmes sélectionnés.
}
\description{
Cette fonction trace un graphique en ligne montrant l'évolution du nombre de sets pour les dix thèmes les plus populaires au fil des années.
}
