% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_5_themes.R
\name{top_5_themes}
\alias{top_5_themes}
\title{Sélectionne les cinq thèmes les plus populaires.}
\usage{
top_5_themes(df_prices)
}
\arguments{
\item{df_prices}{Le dataframe contenant les données sur les prix des sets.}
}
\value{
Un vecteur contenant les noms des cinq thèmes les plus populaires.
}
\description{
Cette fonction sélectionne les cinq thèmes les plus populaires en fonction du nombre de sets dans le dataframe des prix.
}
