% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_themes_barchart.R
\name{top_themes_barchart}
\alias{top_themes_barchart}
\title{Trace un graphique à barres pour les thèmes les plus populaires.}
\usage{
top_themes_barchart(df_prices, themes)
}
\arguments{
\item{df_prices}{Le dataframe contenant les données sur les prix des sets.}

\item{themes}{Un vecteur contenant les noms des thèmes à inclure dans le graphique.}
}
\value{
Un graphique à barres montrant le nombre total de sets pour les thèmes sélectionnés.
}
\description{
Cette fonction trace un graphique à barres montrant le nombre total de sets pour les thèmes sélectionnés.
}
