% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_barchart.R
\name{year_barchart}
\alias{year_barchart}
\title{Trace un graphique à barres pour le nombre de sets par année.}
\usage{
year_barchart(df_prices, years = 2014:2024)
}
\arguments{
\item{df_prices}{Le dataframe contenant les données sur les prix des sets.}

\item{years}{Vecteur contenant les années à inclure dans le graphique. Par défaut, toutes les années de 2014 à 2024 sont incluses.}
}
\value{
Un graphique à barres montrant le nombre total de sets pour chaque année spécifiée.
}
\description{
Cette fonction trace un graphique à barres montrant le nombre total de sets pour chaque année spécifiée.
}
