% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_blocktext.R
\name{year_blocktext}
\alias{year_blocktext}
\title{Génère un texte descriptif pour une année donnée.}
\usage{
year_blocktext(df_prices, year_input)
}
\arguments{
\item{df_prices}{Le dataframe contenant les données sur les prix des sets.}

\item{year_input}{L'année pour laquelle générer le texte descriptif.}
}
\value{
Un texte descriptif indiquant que l'onglet contient des informations sur les sorties LEGO pour l'année spécifiée.
}
\description{
Cette fonction génère un texte descriptif pour une année donnée, indiquant qu'il contient des informations sur les sorties LEGO pour cette année.
}
