% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_mean_pieces.R
\name{year_mean_pieces}
\alias{year_mean_pieces}
\title{Calcule le nombre moyen de pièces pour une année donnée.}
\usage{
year_mean_pieces(df_prices, year_input)
}
\arguments{
\item{df_prices}{Le dataframe contenant les données sur les prix des sets.}

\item{year_input}{L'année pour laquelle calculer le nombre moyen de pièces.}
}
\value{
Le nombre moyen de pièces pour l'année spécifiée.
}
\description{
Cette fonction calcule le nombre moyen de pièces pour une année donnée à partir des données sur les prix des sets LEGO.
}
