% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_mean_price.R
\name{year_mean_price}
\alias{year_mean_price}
\title{Calcule le prix moyen des sets pour une année donnée.}
\usage{
year_mean_price(df_prices, year_input)
}
\arguments{
\item{df_prices}{Le dataframe contenant les données sur les prix des sets.}

\item{year_input}{L'année pour laquelle calculer le prix moyen des sets.}
}
\value{
Le prix moyen des sets pour l'année spécifiée.
}
\description{
Cette fonction calcule le prix moyen des sets pour une année donnée à partir des données sur les prix des sets LEGO.
}
