% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_mean_price_new.R, R/year_price_new.R
\name{year_mean_price_new}
\alias{year_mean_price_new}
\title{Calcule le prix moyen des nouveaux sets pour une année donnée.}
\usage{
year_mean_price_new(df_prices, year_input)

year_mean_price_new(df_prices, year_input)
}
\arguments{
\item{df_prices}{Le dataframe contenant les données sur les prix des sets.}

\item{year_input}{L'année pour laquelle calculer le prix moyen des nouveaux sets.}
}
\value{
Le prix moyen des nouveaux sets pour l'année spécifiée.

Le prix moyen des nouveaux sets pour l'année spécifiée.
}
\description{
Cette fonction calcule le prix moyen des nouveaux sets pour une année donnée à partir des données sur les prix des sets LEGO.

Cette fonction calcule le prix moyen des nouveaux sets pour une année donnée en se basant sur les données de prix des sets.
}
