% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_mean_rating.R
\name{year_mean_rating}
\alias{year_mean_rating}
\title{Calcule la note moyenne des sets pour une année donnée.}
\usage{
year_mean_rating(df_prices, year_input)
}
\arguments{
\item{df_prices}{Le dataframe contenant les données sur les prix des sets.}

\item{year_input}{L'année pour laquelle calculer la note moyenne des sets.}
}
\value{
La note moyenne des sets pour l'année spécifiée.
}
\description{
Cette fonction calcule la note moyenne des sets pour une année donnée à partir des données sur les notes des sets LEGO.
}
