% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_price_multiplier.R
\name{year_price_multiplier}
\alias{year_price_multiplier}
\title{Calcule le multiplicateur de prix pour une année donnée.}
\usage{
year_price_multiplier(df_prices, year_input)
}
\arguments{
\item{df_prices}{Le dataframe contenant les données sur les prix des sets.}

\item{year_input}{L'année pour laquelle calculer le multiplicateur de prix.}
}
\value{
Le multiplicateur de prix pour l'année spécifiée.
}
\description{
Cette fonction calcule le multiplicateur de prix pour une année donnée en comparant le prix moyen des nouveaux sets par rapport au prix moyen des sets de détail pour cette année.
}
\seealso{
\code{\link{year_mean_price}}, \code{\link{year_mean_price_new}}
}
