% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_price_plot.R
\name{year_price_plot}
\alias{year_price_plot}
\title{Trace un graphique en ligne montrant l'évolution du prix de vente moyen des sets au fil des années.}
\usage{
year_price_plot(df_prices)
}
\arguments{
\item{df_prices}{Le dataframe contenant les données sur les prix des sets.}
}
\value{
Un graphique en ligne montrant l'évolution du prix de vente moyen des sets pour chaque année, en fonction de la catégorie.
}
\description{
Cette fonction trace un graphique en ligne montrant l'évolution du prix de vente moyen des sets pour chaque année, en fonction de la catégorie.
}
