% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_price_range.R
\name{year_price_range}
\alias{year_price_range}
\title{Obtient la gamme de prix des sets pour une année donnée.}
\usage{
year_price_range(df_prices, year_input)
}
\arguments{
\item{df_prices}{Le dataframe contenant les données sur les prix des sets.}

\item{year_input}{L'année pour laquelle on souhaite obtenir la gamme de prix des sets.}
}
\value{
La gamme de prix des sets pour l'année spécifiée.
}
\description{
Cette fonction récupère la gamme de prix des sets pour une année donnée à partir des données de prix des sets.
}
