% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_sets_plot.R
\name{year_sets_plot}
\alias{year_sets_plot}
\title{Trace un graphique montrant l'évolution du nombre de sets par année et par catégorie.}
\usage{
year_sets_plot(df_prices)
}
\arguments{
\item{df_prices}{Le dataframe contenant les données sur les prix des sets.}
}
\value{
Un graphique en ligne montrant l'évolution du nombre de sets par année et par catégorie.
}
\description{
Cette fonction trace un graphique en ligne montrant l'évolution du nombre de sets par année et par catégorie.
}
