% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_table.R
\name{year_table}
\alias{year_table}
\title{Tableau interactif avec filtre par année.}
\usage{
year_table(df_prices, year_input)
}
\arguments{
\item{df_prices}{Le dataframe contenant les informations sur les prix des sets LEGO.}

\item{year_input}{L'année(s) à filtrer dans lZe tableau.}
}
\description{
Cette fonction crée un tableau interactif avec un filtre par année à partir du dataframe
fourni. Elle utilise le package gt pour générer une table HTML interactive avec des
fonctionnalités telles que le filtrage, le tri et la mise en forme des données. La table
affiche les informations spécifiques sur les sets LEGO en fonction du thème choisi.
}
