% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_total_sets.R
\name{year_total_sets}
\alias{year_total_sets}
\title{Calcule le nombre total de sets pour une année donnée.}
\usage{
year_total_sets(df_prices, year_input)
}
\arguments{
\item{df_prices}{Le dataframe contenant les données sur les prix des sets.}

\item{year_input}{L'année pour laquelle calculer le nombre total de sets.}
}
\value{
Le nombre total de sets pour l'année spécifiée.
}
\description{
Cette fonction calcule le nombre total de sets sortis pour une année donnée en se basant sur les données de prix des sets.
}
